fix decorator, move to it_utils
