+++ tests/integration/reqs/test_ledger.py
@@ -1,38 +1,10 @@
-import inspect
-
 try:
     from unittest import IsolatedAsyncioTestCase
 except ImportError:
     from aiounittest import AsyncTestCase as IsolatedAsyncioTestCase
 
-from tests.integration.it_utils import ASYNC_JSON_RPC_CLIENT, JSON_RPC_CLIENT
+from tests.integration.it_utils import test_async_and_sync
 from xrpl.models.requests import Ledger
-
-
-def test_async_and_sync(test_function):
-    sync_code = inspect.getsource(test_function)
-    sync_code = (
-        sync_code.replace("async def", "def")  # convert method from async to sync
-        .replace("await ", "")  # replace function calls
-        .replace("@test_async_and_sync\n    ", "")  # avoid decorator recursion
-        # .replace("_async", "") # change methods
-        .replace("\n    ", "\n")  # remove indenting (syntax error otherwise)
-        .replace("    def", "def")  # remove more indenting
-    )
-    first_line = inspect.getsourcelines(test_function)[0][1]
-    sync_code += first_line.replace("    async def ", "").replace(":", "")
-    print(sync_code)
-
-    async def modified_test(self):
-        with self.subTest(version="sync"):
-            exec(sync_code, globals(), {"self": self, "client": JSON_RPC_CLIENT})
-            # NOTE: passing `globals()` into `exec` is really bad practice and not safe
-            # at all, but in this case it's fine because it's only running test code
-            print("success")
-        with self.subTest(version="async"):
-            await test_function(self, ASYNC_JSON_RPC_CLIENT)
-
-    return modified_test
 
 
 class TestLedger(IsolatedAsyncioTestCase):
